+++ src/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2026-02-26
+### Added
+ - WiFi support
+ - Device config overrides
+ - Device runtime flow updata
+
+### Improvements
+ - Updated webserver
+ - Refactored code for improved maintainability
+
+
 ## [Unreleased](https://github.com/gideonmaina/sensors_power_saver/releases/tag/v1.1.0) v.1.1.0 (2025-04-14)
 
 ### Added